[#30] Rename: post 관련 컴포넌트 post 폴더로 이동
