+++ src/components/Board/post/PostCard.tsx
