+++ src/components/Board/post/PostCardContent.tsx
