+++ src/components/Board/post/PostCardHeader.tsx
@@ -1,6 +1,6 @@
 import { Board } from '@/types/board';
 import { formateAgo } from '@/utils/day';
-import UserImage from '../ui/UserImage';
+import UserImage from '../../ui/UserImage';
 
 type Props = Pick<Board, 'name' | 'title' | 'createdAt' | 'userImage'>;
